+++ index.html
@@ -43,6 +43,7 @@
 }
 
 </style>
-<script src="./js.js"></script>
+<script type="module" src="./main.js"></script>
+
  
 </html>
